fix delete

+++ frontend/src/components/Playlist/Playlist.js
@@ -119,10 +119,24 @@ const Playlist = ({ songUrl, setSongUrl }) => {
 
     const handleDelete = (e) => {
         e.preventDefault();
-        // console.log(e.target.id);
-        dispatch(deleteSong(e.target.id));
-        // window.location.reload(false); //too fast
-        setDeletes(deletes+1);
+
+        // console.log(currentUser._id,e.target.id)
+        if(currentUser){
+            if(currentUser._id===e.target.id){
+                // console.log(e.target.id);
+                dispatch(deleteSong(e.target.id));
+                // window.location.reload(false); //too fast
+                setDeletes(deletes+1);
+            } else {
+                
+                return alert('You Must Be The Uploader');
+            }
+        } else {
+            
+            return alert('You Must Be Logged In');
+        }
+
+        
     }
 
     const handleAdd = (e) => {
@@ -140,12 +154,12 @@ const Playlist = ({ songUrl, setSongUrl }) => {
                                 <br></br>
                                 <span className={styles.artistName}>{song.artist}</span>
                         </button>
-                        <p className={styles.addDelete} value={song} id={song._id} onClick={handleDelete}>
+                        <div className={styles.addDelete} value={song} id={song._id} onClick={handleDelete}>
                             {/* <svg id={styles.trashcan}>
                             <path d="M16 6V4.5C16 3.12 14.88 2 13.5 2h-3C9.11 2 8 3.12 8 4.5V6H3v2h1.06l.81 11.21C4.98 20.78 6.28 22 7.86 22h8.27c1.58 0 2.88-1.22 3-2.79L19.93 8H21V6h-5zm-6-1.5c0-.28.22-.5.5-.5h3c.27 0 .5.22.5.5V6h-4V4.5zm7.13 14.57c-.04.52-.47.93-1 .93H7.86c-.53 0-.96-.41-1-.93L6.07 8h11.85l-.79 11.07zM9 17v-6h2v6H9zm4 0v-6h2v6h-2z"></path>
                             </svg> */}
-                        <FontAwesomeIcon icon={faTrashCan} size="lg" />
-                        </p>
+                            <FontAwesomeIcon icon={faTrashCan} size="lg"/>
+                        </div>
                         <AddSongToPlaylistMenu song={song}/>
                     </div>
                 </li>
